Feat: Landing Poprawki buttonów, poprawka readme.
